requirements.txt
Flask
gunicorn
selenium
webdriver-manager
mysql-connector-python
pandas
plotly
